Split autodoc loader functions to a new module
